+++ R/learner_randomForest_regr_randomForest.R
@@ -114,6 +114,7 @@ LearnerRegrRandomForest = R6Class("LearnerRegrRandomForest",
     .train = function(task) {
 
       pars = self$param_set$get_values(tags = "train")
+      pars$weights = private$.get_weights(task)
       formula = task$formula()
       data = task$data()